+++ TPV1 labo/practica 2 que coño es esto/sdl plantilla/HolaSDL/SceneObject.h
@@ -14,7 +14,7 @@ class SceneObject : GameObject
 protected:
 
 	// posicion del objeto
-	Point2D<double> position;
+	Point2D<int> position;
 
 	// dimension del objeto (height & width)
 	int width, height;
@@ -30,7 +30,7 @@ protected:
 	// metodos publicos
 public:
 
-	SceneObject::SceneObject(SDL_Rect destRect, Point2D<double> position, int width, int height, int vidas, Texture* texture, Game* game)
+	SceneObject::SceneObject(SDL_Rect destRect, Point2D<int> position, int width, int height, int vidas, Texture* texture, Game* game)
 		: destRect(destRect), position(position), width(width), height(height), vidas(vidas), texture(texture), GameObject(game)
 	{
 		// ????
